+++ app/Http/Controllers/ConceptoController.php
@@ -91,11 +91,13 @@ class ConceptoController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        Concepto::destroy($id);
-        return redirect()
-            ->route('concepto')
-            ->with('success', 'Eliminado Correctamente');
+        if ($request->ajax()) {
+            Concepto::destroy($id);
+            return response()->json(['mensaje' => 'ok']);
+        } else {
+            abort(404);
+        }
     }
 }
